requirements.txt
streamlit>=1.30
pillow>=10.0
numpy>=1.24
reportlab>=4.0
cairosvg>=2.7
